Ejercicio E/S terminado y entregado

+++ 00-Unidades/01_entrada_salida/Ejercicios_entrada_salida/ES_app_09.py
@@ -33,7 +33,7 @@ class App(customtkinter.CTk):
 
 
     def btn_mostrar_on_click(self):
-        sueldo = int(self.txt_sueldo.get())
+        sueldo = float(self.txt_sueldo.get())
         porcentaje_de_incremento = 15
 
         calculo_del_15 = ((sueldo * porcentaje_de_incremento) // 100) + sueldo
